Function topic_validator::does_valid_topic_match

pub fn does_valid_topic_match(publish: &str, subscribe: &str) -> bool

Checks if a given valid publish topic matches a subscribe topic. Only use if you are certain that a topic name is valid. If you are not certain use does_topic_match.

Arguments

publish - A string slice representing the publish topic.

subscribe - A string slice representing the subscribe topic.

Returns

A boolean value indicating whether the publish topic matches the subscribe topic.

Examples

use topic_validator::does_valid_topic_match;

let publish = "sports/tennis/player1";
let subscribe = "sports/tennis/player1";
assert!(does_valid_topic_match(publish, subscribe));

let publish = "sports/tennis/player1";
let subscribe = "sports/tennis/player2";
assert!(!does_valid_topic_match(publish, subscribe));